Split up entity/dataset model
